+++ config/i18n.json
@@ -347,6 +347,7 @@
     "Inventory": "Inventory",
     "IosPwaPrompt": "In Safari, click the share icon (middle button on the bottom) and select \"Add to Home Screen\".",
     "KeyboardShortcuts": "Keyboard Shortcuts",
+    "MaterialCounts": "Material Counts",
     "Menu": "Menu",
     "Refresh": "Refresh Destiny Data [R]",
     "ReloadApp": "Reload App",
