+++ src/app/loadout-drawer/LoadoutPopup.tsx
@@ -1,4 +1,3 @@
-import { LoadoutSort } from '@destinyitemmanager/dim-api-types';
 import { languageSelector, settingSelector } from 'app/dim-api/selectors';
 import { AlertIcon } from 'app/dim-ui/AlertIcon';
 import ClassIcon from 'app/dim-ui/ClassIcon';
@@ -10,11 +9,13 @@ import { allItemsSelector, bucketsSelector, hasClassifiedSelector } from 'app/in
 import { editLoadout } from 'app/loadout-drawer/loadout-events';
 import MaxlightButton from 'app/loadout-drawer/MaxlightButton';
 import { useDefinitions } from 'app/manifest/selectors';
+import { showMaterialCount } from 'app/material-counts/MaterialCountsWrappers';
 import { ItemFilter } from 'app/search/filter-types';
 import { plainString } from 'app/search/search-filters/freeform';
 import { RootState, ThunkDispatchProp } from 'app/store/types';
 import { isiOSBrowser } from 'app/utils/browsers';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
+import consumablesIcon from 'destiny-icons/general/consumables.svg';
 import _ from 'lodash';
 import React, { useState } from 'react';
 import { connect, useSelector } from 'react-redux';
@@ -84,7 +85,7 @@ function mapStateToProps() {
             loadout.classType === DestinyClass.Unknown ||
             loadout.classType === dimStore.classType
         ),
-        loadoutSort === LoadoutSort.ByEditTime ? (l) => -(l.lastUpdatedAt ?? 0) : (l) => l.name
+        loadoutSort === 0 ? (l) => -(l.lastUpdatedAt ?? 0) : (l) => l.name
       )
   );
 
@@ -224,6 +225,15 @@ function LoadoutPopup({
         </li>
       )}
       <ul className={styles.list}>
+        {!filteringLoadouts && dimStore.isVault && isPhonePortrait && (
+          <li className={styles.menuItem}>
+            <span onClick={showMaterialCount}>
+              <img src={consumablesIcon} />
+              <span>{t('Header.MaterialCounts')}</span>
+            </span>
+          </li>
+        )}
+
         {!filteringLoadouts && query.length > 0 && (
           <li className={styles.menuItem}>
             <span onClick={applySearchLoadout}>
