+++ src/locale/en.json
@@ -321,6 +321,7 @@
     "IosPwaPrompt": "In Safari, click the share icon (middle button on the bottom) and select \"Add to Home Screen\".",
     "KeyboardShortcuts": "Keyboard Shortcuts",
     "LaunchDIMAlone": "Separate Window",
+    "MaterialCounts": "Material Counts",
     "Menu": "Menu",
     "Refresh": "Refresh Destiny Data [R]",
     "ReloadApp": "Reload App",
